refactor: separated router function and controller logic
